backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
drf-yasg==1.21.7
Pillow==10.1.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
dj-rest-auth==5.0.1
channels==4.0.0
channels-redis==4.1.0
django-storages==1.14.2
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0 